refactor: 클래스 위치 및 패키지 변경
