+++ growthmate-batch/src/main/java/com/growup/growthmate/batch/company/XSSFRowToCompanyMapper.java
@@ -1,8 +1,8 @@
 package com.growup.growthmate.batch.company;
 
 import com.growup.growthmate.company.domain.Company;
-import com.growup.growthmate.support.XSSFRowMapper;
-import com.growup.growthmate.support.XSSFRowUtils;
+import com.growup.growthmate.support.exel.XSSFRowMapper;
+import com.growup.growthmate.support.exel.XSSFRowUtils;
 import org.apache.poi.xssf.usermodel.XSSFRow;
 import org.springframework.stereotype.Component;
