+++ growthmate-batch/src/main/java/com/growup/growthmate/support/exel/XSSFRowMapper.java
@@ -1,4 +1,4 @@
-package com.growup.growthmate.support;
+package com.growup.growthmate.support.exel;
 
 import org.apache.poi.xssf.usermodel.XSSFRow;
