+++ growthmate-batch/src/test/java/com/growup/growthmate/batch/CompanyBatchTest.java
@@ -3,13 +3,16 @@ package com.growup.growthmate.batch;
 import com.growup.growthmate.isolation.TestIsolation;
 import org.junit.jupiter.api.Test;
 import org.springframework.batch.core.BatchStatus;
+import org.springframework.batch.core.ExitStatus;
 import org.springframework.batch.core.JobExecution;
+import org.springframework.batch.core.StepExecution;
 import org.springframework.batch.test.JobLauncherTestUtils;
 import org.springframework.batch.test.context.SpringBatchTest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertAll;
 
 @SpringBootTest
 @SpringBatchTest
@@ -25,6 +28,11 @@ class CompanyBatchTest {
         JobExecution actual = jobLauncherTestUtils.launchJob();
 
         // then
-        assertThat(actual.getStatus()).isEqualTo(BatchStatus.COMPLETED);
+        assertAll(
+                () -> assertThat(actual.getStatus()).isEqualTo(BatchStatus.COMPLETED),
+                () -> assertThat(actual.getStepExecutions())
+                        .map(StepExecution::getExitStatus)
+                        .containsOnly(ExitStatus.COMPLETED)
+        );
     }
 }
